+++ CHANGELOG.md
@@ -5,6 +5,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Dates in this file meets Gregorian calendar. Date in format YYYY-MM-DD.
 
+## [2.0.4] - [2025-07-22], PerseusDarkMode
+
+### Changed
+
+- Access modifier of DARK_MODE_SETTINGS_KEY constant.
+
+### Removed
+
+- Platforms specification of Package manifesto. 
+
+### Updated
+
+- GitHub CI scripts.
+- CPL dependency to v1.5.1.
+
+## [2.0.3] - [2025-06-18], PerseusDarkMode
+
+### Fixed
+
+- Logger access modifier, typealias `log` must be not public.
+
 ## [2.0.2] - [2025-06-17], PerseusDarkMode
 
 ### Upgraded